+++ config/config.php
@@ -12,10 +12,14 @@ declare(strict_types=1);
 return [
 
     'app' => [
-        'urlProtocol'       => 'http://',
-        'urlSubFolder'      => '/app', // es /var/www/html/app/
-        'urlPublicFolder'   => '/app/public', // es /var/www/html/app/public
-        'useCompiledRoutes' => false
+        //protocol utilized [http://, https://]
+        'protocol'       => 'http://',
+        //folder of the app, if app isn't in the web server root
+        'subFolder'      => '/app',
+        //public folder of the app
+        'publicFolder'   => '/app/public',
+        //define if app use routes exported in static array [true, false]
+        'compiledRoutes' => false
     ],
 
     'session' => [